# ComplianceService/Models/IdentityCardCheckResult.cs
/*
 * Сервис проверки контрагентов с целью ПОД/ФТ
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.2-oas3
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace ComplianceService.Models
{ 
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public partial class IdentityCardCheckResult : IndividualParams, IEquatable<IdentityCardCheckResult>
    { 
        /// <summary>
        /// Результат проверки ДУЛ
        /// </summary>
        /// <value>Результат проверки ДУЛ</value>
        [DataMember(Name="checkResult")]
        public string CheckResult { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class IdentityCardCheckResult {\n");
            sb.Append("  CheckResult: ").Append(CheckResult).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public  new string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (ReferenceEquals(null, obj)) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((IdentityCardCheckResult)obj);
        }

        /// <summary>
        /// Returns true if IdentityCardCheckResult instances are equal
        /// </summary>
        /// <param name="other">Instance of IdentityCardCheckResult to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(IdentityCardCheckResult other)
        {
            if (ReferenceEquals(null, other)) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    CheckResult == other.CheckResult ||
                    CheckResult != null &&
                    CheckResult.Equals(other.CheckResult)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    if (CheckResult != null)
                    hashCode = hashCode * 59 + CheckResult.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(IdentityCardCheckResult left, IdentityCardCheckResult right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(IdentityCardCheckResult left, IdentityCardCheckResult right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}